Add star of evaluation Relates #34
